+++ src/nft/components/details/detailsV2/DataPageHeader.tsx
@@ -1,12 +1,11 @@
-import { Trans } from '@lingui/macro'
-import { ButtonGray, ButtonPrimary } from 'components/Button'
 import Column from 'components/Column'
 import Row from 'components/Row'
-import { HandHoldingDollarIcon, VerifiedIcon } from 'nft/components/icons'
+import { VerifiedIcon } from 'nft/components/icons'
 import { GenieAsset } from 'nft/types'
-import { formatEth } from 'nft/utils'
-import styled, { css } from 'styled-components/macro'
+import styled from 'styled-components/macro'
 import { BREAKPOINTS, ThemedText } from 'theme'
+
+import { BuyButton } from './BuyButton'
 
 const HeaderContainer = styled(Row)`
   gap: 24px;
@@ -28,38 +27,7 @@ const AssetText = styled(Column)`
   margin-right: auto;
 `
 
-const ButtonStyles = css`
-  width: min-content;
-  flex-shrink: 0;
-  border-radius: 16px;
-`
-
-const BuyButton = styled(ButtonPrimary)`
-  display: flex;
-  flex-direction: row;
-  padding: 16px 24px;
-  gap: 8px;
-  line-height: 24px;
-  white-space: nowrap;
-  ${ButtonStyles}
-`
-
-const Price = styled.div`
-  color: ${({ theme }) => theme.accentTextLightSecondary};
-`
-
-const MakeOfferButtonSmall = styled(ButtonPrimary)`
-  padding: 16px;
-  ${ButtonStyles}
-`
-
-const MakeOfferButtonLarge = styled(ButtonGray)`
-  white-space: nowrap;
-  ${ButtonStyles}
-`
-
 export const DataPageHeader = ({ asset }: { asset: GenieAsset }) => {
-  const price = asset.sellorders?.[0]?.price.value
   return (
     <HeaderContainer>
       <AssetImage src={asset.imageUrl} />
@@ -73,21 +41,7 @@ export const DataPageHeader = ({ asset }: { asset: GenieAsset }) => {
         </ThemedText.HeadlineMedium>
       </AssetText>
       <Row justifySelf="flex-end" width="min-content" gap="12px">
-        {price ? (
-          <>
-            <BuyButton>
-              <Trans>Buy</Trans>
-              <Price>{formatEth(price)} ETH</Price>
-            </BuyButton>
-            <MakeOfferButtonSmall>
-              <HandHoldingDollarIcon />
-            </MakeOfferButtonSmall>
-          </>
-        ) : (
-          <MakeOfferButtonLarge>
-            <Trans>Make an offer</Trans>
-          </MakeOfferButtonLarge>
-        )}
+        <BuyButton asset={asset} onDataPage />
       </Row>
     </HeaderContainer>
   )
