+++ src/nft/components/details/detailsV2/LandingPage.tsx
@@ -1,12 +1,12 @@
 import Column, { ColumnCenter } from 'components/Column'
 import Row from 'components/Row'
 import { VerifiedIcon } from 'nft/components/icons'
-import { useFetchSingleAsset } from 'nft/hooks/useFetchAssets'
 import { CollectionInfoForAsset, GenieAsset } from 'nft/types'
 import { useEffect, useRef } from 'react'
 import styled from 'styled-components/macro'
 import { BREAKPOINTS } from 'theme'
 
+import { BuyButton } from './BuyButton'
 import { InfoChips } from './InfoChips'
 import { MediaRenderer } from './MediaRenderer'
 
@@ -108,8 +108,6 @@ interface LandingPageProps {
 }
 
 export const LandingPage = ({ asset, collection, setShowDataHeader }: LandingPageProps) => {
-  const fetchAndPurchaseAsset = useFetchSingleAsset()
-
   const intersectionRef = useRef<HTMLDivElement>(null)
   const observableRef = useRef(
     new IntersectionObserver((entries) => {
@@ -147,7 +145,7 @@ export const LandingPage = ({ asset, collection, setShowDataHeader }: LandingPag
           <StyledHeadlineText>{asset.name ?? `${asset.collectionName} #${asset.tokenId}`}</StyledHeadlineText>
         </InfoDetailsContainer>
         <InfoChips asset={asset} />
-        <button onClick={() => fetchAndPurchaseAsset(asset)}>Buy now</button>
+        <BuyButton asset={asset} />
       </InfoContainer>
     </LandingPageContainer>
   )
