+++ src/nft/components/details/detailsV2/shared.ts
@@ -34,3 +34,9 @@ export const Scrim = styled.div<{ isBottom?: boolean }>`
     `linear-gradient(180deg, ${opacify(0, theme.backgroundSurface)} 0%, ${theme.backgroundSurface} 100%)`};
   display: flex;
 `
+
+const ButtonStyles = css`
+  width: min-content;
+  flex-shrink: 0;
+  border-radius: 16px;
+`
